+++ src/refactoring/3_widgets/product-list/types.ts
@@ -2,9 +2,6 @@ import { Product, Discount } from "../../6_shared/types";
 
 export interface ProductCardProps {
   product: Product;
-  remainingStock: number;
-  addToCart: (product: Product) => void;
-  maxDiscount: number;
 }
 
 export interface DiscountListProps {
