+++ src/refactoring/4_features/cart/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from "./use-cart";
-export * from "./use-cart-with-coupon";
