--- src/refactoring/4_features/cart/hooks/use-cart-with-coupon.ts
@@ -1,19 +1,0 @@
-import { useAtomValue } from "jotai";
-
-import { cartTotalAtom } from "../../../5_entities/cart/model";
-import { calculations } from "../../../5_entities/cart/lib";
-import { selectedCouponAtom } from "../../../5_entities/coupon/model";
-
-export const useCartWithCoupon = () => {
-  const cartTotal = useAtomValue(cartTotalAtom);
-  const selectedCoupon = useAtomValue(selectedCouponAtom);
-
-  const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } =
-    calculations.calculateCouponAppliedTotal(cartTotal, selectedCoupon);
-
-  return {
-    totalBeforeDiscount,
-    totalAfterDiscount,
-    totalDiscount,
-  };
-};
