+++ src/refactoring/4_features/order-summary/hooks/use-order-summary.ts
@@ -1,0 +1,18 @@
+import { useAtomValue } from "jotai";
+
+import { orderSummaryAtom } from "../model";
+import { calculateCouponAppliedTotal } from "../lib";
+import { CartTotal } from "../../../6_shared/types";
+
+export const useOrderSummary = (): CartTotal => {
+  const { cartTotal, selectedCoupon } = useAtomValue(orderSummaryAtom);
+
+  const { totalBeforeDiscount, totalAfterDiscount, totalDiscount } =
+    calculateCouponAppliedTotal(cartTotal, selectedCoupon);
+
+  return {
+    totalBeforeDiscount,
+    totalAfterDiscount,
+    totalDiscount,
+  };
+};
