+++ src/refactoring/4_features/order-summary/model/index.ts
@@ -1,0 +1,1 @@
+export * from "./atoms";
