+++ src/refactoring/4_features/product/hooks/use-products.ts
@@ -1,7 +1,6 @@
 import { useAtom } from "jotai";
 
 import { productsAtom } from "../../../5_entities/product/model/atoms";
-import { getMaxDiscount } from "../../../5_entities/product/lib";
 
 export const useProducts = () => {
   const [products, setProducts] = useAtom(productsAtom);
@@ -9,6 +8,5 @@ export const useProducts = () => {
   return {
     products,
     setProducts,
-    getMaxDiscount,
   };
 };
